functions/load/requirements.txt
google-cloud-bigquery==3.12.0